+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+-  One Webtools event id published to multiple calendars creates duplicate events [#155](https://github.com/rokwire/gateway-building-block/issues/155)
 
 ## [2.17.3] - 2025-07-23
 ### Changed